Implemented product form component
